Refactor: Change file paths
